+++ oneflow/api/python/framework/tensor.cpp
@@ -185,61 +185,6 @@ static PyObject* PyTensorObject_pin_memory(PyObject* self, PyObject* unused) {
   END_HANDLE_ERRORS
 }
 
-int64_t unpack_long(PyObject* self) {
-  int overflow = -1;
-  long long val = PyLong_AsLongLongAndOverflow(self, &overflow);
-  if (val == -1 && PyErr_Occurred()) {
-    THROW(RuntimeError) << "unpack_long >> Python exception occurs. python type:"
-                        << Py_TYPE(self)->tp_name;
-  }
-  if (overflow != 0) { THROW(RuntimeError) << "unpack_long >> Overflow when unpacking long"; }
-  return (int64_t)val;
-}
-
-static PyObject* PyTensorObject_reshape(PyObject* self, PyObject* args) {
-  HANDLE_ERRORS
-  auto tensor = PyTensor_Unpack(self);
-  if (!PyTuple_Check(args)) {
-    THROW(TypeError) << "reshape(): argument 'shape' must be tuple of ints, but found "
-                     << Py_TYPE(args)->tp_name;
-  }
-
-  size_t size = (size_t)PyTuple_Size(args);
-  if (size == -1) { size = tensor->ndim(); }
-  PyObject* shape = PyTuple_GetItem(args, 0);
-  if (PyList_Check(shape)) {
-    size = (size_t)PyList_Size(shape);
-    DimVector vec(size);
-    for (int i = 0; i < size; ++i) { vec.at(i) = unpack_long(PyList_GetItem(shape, i)); }
-    auto result = CHECK_JUST(functional::Reshape(tensor, Shape(vec)));
-    return PyTensor_New(result);
-
-  } else {
-    if (PyLong_Check(shape)) {
-      DimVector vec(size);
-      for (int i = 0; i < size; ++i) { vec.at(i) = unpack_long(PyTuple_GetItem(args, i)); }
-      auto result = CHECK_JUST(functional::Reshape(tensor, Shape(vec)));
-      return PyTensor_New(result);
-    } else {
-      size = (size_t)PyTuple_Size(shape);
-      DimVector vec(size);
-      for (int i = 0; i < size; ++i) { vec.at(i) = unpack_long(PyTuple_GetItem(shape, i)); }
-      auto result = CHECK_JUST(functional::Reshape(tensor, Shape(vec)));
-      return PyTensor_New(result);
-    }
-  }
-  END_HANDLE_ERRORS
-}
-
-static PyObject* PyTensorObject_reshape_as(PyObject* self, PyObject* args) {
-  HANDLE_ERRORS
-  auto tensor = PyTensor_Unpack(self);
-  PyObject* other = PyTuple_GetItem(args, 0);
-  auto result = CHECK_JUST(functional::Reshape(tensor, *PyTensor_Unpack(other)->shape()));
-  return PyTensor_New(result);
-  END_HANDLE_ERRORS
-}
-
 static PyObject* PyTensorObject_requires_grad_(PyObject* self, PyObject* args, PyObject* kwargs) {
   HANDLE_ERRORS
   int requires_grad = 1;
@@ -418,13 +363,26 @@ static PyObject* PyTensorObject__register_storage_delete_hook(PyObject* self, Py
   END_HANDLE_ERRORS
 }
 
+static std::vector<PyMethodDef> concat_method_def(PyMethodDef methods[],
+                                                  PyMethodDef extra_methods[]) {
+  int len1 = 0;
+  int len2 = 0;
+  PyMethodDef* p1 = methods;
+  PyMethodDef* p2 = extra_methods;
+  while ((p1++)->ml_name != NULL) { len1++; }
+  while ((p2++)->ml_name != NULL) { len2++; }
+  std::vector<PyMethodDef> total_methods(len1 + len2 + 1);
+  for (int i = 0; i < len1; i++) total_methods[i] = methods[i];
+  for (int i = 0; i < len2; i++) total_methods[i + len1] = extra_methods[i];
+  total_methods[len1 + len2] = {NULL};
+  return total_methods;
+}
+
 static PyMethodDef PyTensorObject_methods[] = {
     {"storage_offset", PyTensorObject_storage_offset, METH_NOARGS, NULL},
     {"stride", PyTensorObject_stride, METH_NOARGS, NULL},
     {"is_contiguous", PyTensorObject_is_contiguous, METH_NOARGS, NULL},
     {"contiguous", PyTensorObject_contiguous, METH_NOARGS, NULL},
-    {"reshape", PyTensorObject_reshape, METH_VARARGS, NULL},
-    {"reshape_as", PyTensorObject_reshape_as, METH_VARARGS, NULL},
     {"pin_memory", PyTensorObject_pin_memory, METH_NOARGS, NULL},
     {"requires_grad_", (PyCFunction)PyTensorObject_requires_grad_, METH_VARARGS | METH_KEYWORDS,
      NULL},
@@ -632,6 +590,8 @@ static PyHeapTypeObject* MakeTensorMetaclass() {
   return heap_type;
 }
 
+extern PyMethodDef PyTensorObject_extra_methods[];
+
 static PyHeapTypeObject* TensorMetaclass_Type = MakeTensorMetaclass();
 
 static PyTypeObject* MakeTensorType() {
@@ -649,9 +609,13 @@ static PyTypeObject* MakeTensorType() {
   type->tp_init = PyTensorObject_init;
   type->tp_dealloc = PyTensorObject_dealloc;
   type->tp_getset = PyTensorObject_properties;
-  type->tp_methods = PyTensorObject_methods;
-
   type->tp_as_number = &heap_type->as_number;
+
+  static std::vector<PyMethodDef> total_methods =
+      concat_method_def(PyTensorObject_methods, PyTensorObject_extra_methods);
+  type->tp_methods = total_methods.data();
+
+
   type->tp_as_sequence = &PyTensorObject_as_sequence;
   type->tp_as_mapping = &PyTensorObject_as_mapping;
